// src/app/interfaces/product.ts
export interface Product {
    id: string
    Category: string
    MainCategory: string
    TaxTarifCode: string
    SupplierName: string
    WeightMeasure: number
    WeightUnit: string
    Description: string
    Name: string
    DateOfSale?: string
    ProductPicUrl: string
    Status: string
    Quantity: number
    UoM: string
    CurrencyCode: string
    Price: number
    Width: number
    Depth: number
    Height: number
    DimUnit: string
}